requirements.txt
youtube-transcript-api==0.6.3
sentence-transformers==3.4.1
chromadb==0.6.3
google-generativeai==0.8.5
langchain-text-splitters==0.3.5
diskcache==5.6.3
python-dotenv==1.0.1
loguru==0.7.3
